requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jplephem==2.22
MarkupSafe==2.1.5
numpy==1.26.4
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sgp4==2.23
skyfield==1.48
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.3
